refactor: replace exceptions with result classes (#44)
* use HttpResult in PatientController

* use HttpResult in DoctorController

* use HttpResult in RefreshTokenController

* use HttpResult in PrescriptionController

* use HttpResult in AppointmentController

* use CommonResult in AppointmentService

* remove blank line

* remove throwing exceptions from refresh token service

* remove throwing exceptions from repositories

* replace [0] index with First() method

* remove redundant checks in accessing jwt fields

* inherit from UserManager

* inherit from RoleManager

* fix everything but tests

* fix typo in appsettings

* make tests compile (not pass yet)

* move test cases to appUserManager tests (not passing yet)

* fix some of the unit tests

* fix more tests

* fix validate password async unit tests

* fix get user claims tests

* fix remaining tests

* remove getting role claims in GetUserClaims

* add WithFieldError option for custom results

* uncomment validator rule

* remove GetUserByRefreshTokenAsync method

+++ DoctorsOffice.Application/CQRS/Commands/Authenticate/AuthenticateHandler.cs
@@ -1,47 +1,51 @@
-﻿using DoctorsOffice.Application.Services.Auth;
-using DoctorsOffice.Application.Services.Jwt;
-using DoctorsOffice.Application.Services.User;
+﻿using DoctorsOffice.Application.Services.Jwt;
+using DoctorsOffice.Application.Services.RefreshTokens;
 using DoctorsOffice.Domain.DTO.Responses;
-using DoctorsOffice.Domain.Entities.UserTypes;
 using DoctorsOffice.Domain.Utils;
+using DoctorsOffice.Infrastructure.Identity;
 using MediatR;
-using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Http;
 
 namespace DoctorsOffice.Application.CQRS.Commands.Authenticate;
 
 public class AuthenticateHandler : IRequestHandler<AuthenticateCommand, HttpResult<AuthenticateResponse>>
 {
+    private readonly AppUserManager _appUserManager;
     private readonly IJwtService _jwtService;
     private readonly IRefreshTokenService _refreshTokenService;
-    private readonly UserManager<AppUser> _userManager;
-    private readonly IUserService _userService;
 
     public AuthenticateHandler(
-        IUserService userService,
         IJwtService jwtService,
         IRefreshTokenService refreshTokenService,
-        UserManager<AppUser> userManager)
+        AppUserManager appUserManager)
     {
-        _userService = userService;
         _jwtService = jwtService;
         _refreshTokenService = refreshTokenService;
-        _userManager = userManager;
+        _appUserManager = appUserManager;
     }
 
-    public async Task<HttpResult<AuthenticateResponse>> Handle(AuthenticateCommand request,
-        CancellationToken cancellationToken)
+    public async Task<HttpResult<AuthenticateResponse>> Handle(
+        AuthenticateCommand request, CancellationToken cancellationToken)
     {
         var result = new HttpResult<AuthenticateResponse>();
 
-        var user = await _userService.GetUserByUserNameAsync(request.UserName);
-        var userClaims = await _userService.GetUserRolesAsClaimsAsync(user);
+        var findByNameResult = await _appUserManager.FindByNameAsync(request.UserName);
+        if (findByNameResult.IsFailed || findByNameResult.Value is null)
+        {
+            return result
+                .WithError(findByNameResult.Error)
+                .WithStatusCode(StatusCodes.Status404NotFound);
+        }
+
+        var user = findByNameResult.Value;
+        var userClaims = await _jwtService.GetUserClaimsAsync(user);
         var jwtToken = _jwtService.GenerateJwtToken(userClaims);
-        var refreshToken = await _jwtService.GenerateRefreshTokenAsync(request.IpAddress, cancellationToken);
+        var refreshToken = await _refreshTokenService.GenerateRefreshTokenAsync(request.IpAddress, cancellationToken);
 
         _refreshTokenService.RemoveOldRefreshTokens(user);
 
         user.RefreshTokens.Add(refreshToken);
-        await _userManager.UpdateAsync(user);
+        await _appUserManager.UpdateAsync(user);
 
         return result.WithValue(new AuthenticateResponse
         {
